requirements.txt
aiofiles==0.3.2
aiohttp==3.3.2
bottom==2.1.3
colorama==0.3.9
colorhash==1.0.2
discord.py==0.16.12
python-dateutil==2.7.3
pytz==2018.4
toml==0.9.4
unit-converter==1.1.0
